use a viewModel to provide the locationsource
